requirements.txt
beautifulsoup4==4.5.1
python-dateutil==2.6.0
readchar==0.7
requests==2.12.4
six==1.10.0
